Fix CoreRT build breaks after Enum move to shared partition
